+++ src/AppView.tsx
@@ -1,34 +1,28 @@
 // import { MemoryRouter as Router, Routes, Route } from 'react-router-dom';
-import React, { useEffect, useRef, WheelEvent } from 'react';
+import React, { useEffect, useRef } from 'react';
 import { observer } from 'mobx-react-lite';
 
 import {
-  Alert,
   Box,
   Button,
   ButtonPropsColorOverrides,
   FormControl,
   FormControlLabel,
-  IconButton,
   InputLabel,
   MenuItem,
   Select,
-  Slide,
-  Snackbar,
   Switch,
   Tooltip,
   Typography,
 } from '@mui/material';
 import { OverridableStringUnion } from '@mui/types';
 import { ThemeProvider, createTheme } from '@mui/material/styles';
-import ArrowDownwardIcon from '@mui/icons-material/ArrowDownward';
 import ClearIcon from '@mui/icons-material/Clear';
 import SettingsIcon from '@mui/icons-material/Settings';
 import CssBaseline from '@mui/material/CssBaseline';
-import { SnackbarProvider, enqueueSnackbar } from 'notistack';
+import { SnackbarProvider } from 'notistack';
 
 import { App, PortState, portStateToButtonProps } from './model/App';
-import { StatusMsg, StatusMsgSeverity } from './model/StatusMsg';
 import './App.css';
 import {
   DataViewConfiguration,
@@ -37,6 +31,7 @@ import {
 import SettingsDialog from './Settings/SettingsView';
 import TerminalView from './TerminalView';
 import LogoImage from './logo192.png';
+import styles from './AppView.module.css'
 
 import KofiButton from "kofi-button"
 
@@ -73,6 +68,14 @@ const darkTheme = createTheme({
   },
 });
 
+/**
+ * Maps a port state to a colour used in the port state status background on the bottom toolbar.
+ */
+const portStateToBackgroundColor: { [key in PortState]: string; } = {
+  [PortState.CLOSED]: 'red',
+  [PortState.OPENED]: 'green',
+};
+
 interface Props {
   app: App;
 }
@@ -87,48 +90,21 @@ const AppView = observer((props: Props) => {
     }
   }, [app.statusMsgs.length, app.statusMsgScrollLock]);
 
-  // Generate UI showing the status messages
-  const statusMsgs = app.statusMsgs.map((statusMsg: StatusMsg) => {
-    if (statusMsg.severity === StatusMsgSeverity.INFO) {
-      return (
-        <span key={statusMsg.id} style={{ display: 'block' }}>
-          {statusMsg.msg}
-        </span>
-      );
-      // eslint-disable-next-line no-else-return
-    } else if (statusMsg.severity === StatusMsgSeverity.OK) {
-      return (
-        <span key={statusMsg.id} style={{ display: 'block', color: 'green' }}>
-          {statusMsg.msg}
-        </span>
-      );
-      // eslint-disable-next-line no-else-return
-    } else if (statusMsg.severity === StatusMsgSeverity.ERROR) {
-      return (
-        <span key={statusMsg.id} style={{ display: 'block', color: 'red' }}>
-          ERROR: {statusMsg.msg}
-        </span>
-      );
-    } else {
-      throw Error('Unrecognized severity.');
-    }
-  });
-
   // TERMINAL CREATION
   // =================
   // Create terminals based on selected configuration
   let terminals;
   if (app.settings.dataProcessing.appliedData.fields.dataViewConfiguration.value === DataViewConfiguration.SINGLE_TERMINAL) {
     // Show only 1 terminal
-    terminals = <TerminalView appStore={app} terminal={app.txRxTerminal} />;
+    terminals = <TerminalView appStore={app} terminal={app.txRxTerminal} testId='tx-rx-terminal-view' />;
   } else if (app.settings.dataProcessing.appliedData.fields.dataViewConfiguration.value === DataViewConfiguration.SEPARATE_TX_RX_TERMINALS) {
     // Shows 2 terminals, 1 for TX data and 1 for RX data
     terminals = <div style={{ flexGrow: 1, height: '100%', display: 'flex', flexDirection: 'column' }}>
       <div style={{ height: '50%', display: 'flex' }}>
-        <TerminalView appStore={app} terminal={app.txTerminal} />
+        <TerminalView appStore={app} terminal={app.txTerminal} testId='tx-terminal-view'/>
       </div>
       <div style={{ height: '50%', display: 'flex' }}>
-      <TerminalView appStore={app} terminal={app.rxTerminal} />
+        <TerminalView appStore={app} terminal={app.rxTerminal} testId='rx-terminal-view' />
       </div>
     </div>;
   } else {
@@ -179,9 +155,11 @@ const AppView = observer((props: Props) => {
               alignItems: 'center',
               height: '40px',
               gap: '10px',
-              marginBottom: '10px',
-            }}>
+              marginBottom: '10px'}}>
+            {/* ================== LOGO ==================== */}
             <img src={LogoImage} alt="NinjaTerm logo." style={{ width: '30px' }} />
+
+            {/* ================== SETTINGS BUTTON ==================== */}
             <Button
               variant="outlined"
               onClick={() => {
@@ -191,6 +169,7 @@ const AppView = observer((props: Props) => {
               data-testid="settings-button">
               Settings
             </Button>
+            {/* ================== OPEN/CLOSE BUTTON ==================== */}
             <Button
               variant="outlined"
               color={
@@ -219,7 +198,7 @@ const AppView = observer((props: Props) => {
               }}
               startIcon={portStateToButtonProps[app.portState].icon}
               disabled={app.port === null}
-            >
+              sx={{ width: '150px' }}> {/* Specify a width to prevent it resizing when the text changes */}
               {portStateToButtonProps[app.portState].text}
             </Button>
             {/* ================== CLEAR DATA BUTTON ==================== */}
@@ -297,14 +276,20 @@ const AppView = observer((props: Props) => {
               display: 'flex',
               flexDirection: 'row',
               justifyContent: 'end',
+              alignItems: 'center',
               fontSize: '0.9rem',
               gap: '20px',
             }}
             style={{ height: '20px' }}
           >
+            {/* TX/RX ACTIVITY INDICATORS */}
+            {/* Use the key prop here to make React consider this a new element everytime the number
+            of bytes changes. This will re-trigger the flashing animation as desired */}
+            <Box key={app.numBytesTransmitted} className={styles.ledblue}>TX</Box>
+            <Box key={app.numBytesReceived} className={styles.ledyellow}>RX</Box>
             {/* Show port configuration in short hand, e.g. "115200 8n1" */}
             <Box>{app.settings.shortSerialConfigName}</Box>
-            <Box>Port {PortState[app.portState]}</Box>
+            <Box sx={{ backgroundColor: portStateToBackgroundColor[app.portState], padding: '0 10px' }}>Port {PortState[app.portState]}</Box>
           </Box>
         </div>
         {/* The SnackBar's position in the DOM does not matter, it is not positioned in the doc flow.