+++ CHANGELOG.md
@@ -7,6 +7,24 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## [4.2.0] - 2023-09-27
+
+### Added
+
+- Testing function which sets up the App with a mocked WebSerial API is working.
+- Error message is now shown if user attempts to open already in-use port, closes #254.
+- Smart scrolling when max. scrollback size is reached, auto-scroll occurs to keep the same data in view. 
+- Added specific error handling for a BufferOverrunError.
+- Easy to identify colour to the background of the port state in bottom statusbar.
+- Added a margin between the outer edges of each terminal and the displayed text inside it.
+- Flashing indicators added to the bottom status bar to indicate when TX or RX data is received.
+
+### Fixed
+
+- Test which checks written data is working again.
+- Max. scrollback buffer size value now obeyed.
+- Non-fatal errors such as a BufferOverrunError or BreakError keep the connection open.
+
 ## [4.1.0] - 2023-09-24
 
 ### Added